feat: share zod between ui + be
