+++ ladderly-io/src/server/api/routers/user.ts
@@ -1,11 +1,6 @@
 // src/server/api/routers/user.ts
 
-import {
-  PaymentTierEnum,
-  Prisma,
-  type Subscription,
-  type User,
-} from '@prisma/client'
+import { PaymentTierEnum, Prisma } from '@prisma/client'
 import { TRPCError } from '@trpc/server'
 import { z } from 'zod'
 import {
@@ -14,73 +9,14 @@ import {
   publicProcedure,
 } from '~/server/api/trpc'
 import { NULL_RESULT_TRPC_INT } from '~/server/constants'
-
-const tiersOrder = {
-  FREE: 0,
-  PAY_WHAT_YOU_CAN: 1,
-  PREMIUM: 2,
-} as const
-
-// Define the settings input type
-export const UpdateUserSettingsSchema = z.object({
-  email: z.string(),
-  emailBackup: z.string().optional(),
-  emailStripe: z.string().optional(),
-  hasInPersonEventInterest: z.boolean(),
-  hasLiveStreamInterest: z.boolean(),
-  hasOnlineEventInterest: z.boolean(),
-  hasOpenToRelocation: z.boolean(),
-  hasOpenToWork: z.boolean(),
-  hasPublicProfileEnabled: z.boolean(),
-  hasShoutOutsEnabled: z.boolean(),
-  hasSmallGroupInterest: z.boolean(),
-  nameFirst: z.string().nullable(),
-  nameLast: z.string().nullable(),
-  profileBlurb: z.string().nullable(),
-  profileContactEmail: z.string().nullable(),
-  profileCurrentJobCompany: z.string().optional(),
-  profileCurrentJobTitle: z.string().optional(),
-  profileGitHubUri: z.string().nullable(),
-  profileHighestDegree: z.string().optional(),
-  profileHomepageUri: z.string().nullable(),
-  profileLinkedInUri: z.string().nullable(),
-  profileTopNetworkingReasons: z.array(z.string()).optional(),
-  profileTopSkills: z.array(z.string()).optional(),
-  profileYearsOfExperience: z.number().optional(),
-  residenceCountry: z.string(),
-  residenceUSState: z.string(),
-})
-
-// Define the settings output type
-export const GetUserSettingsSchema = UpdateUserSettingsSchema.extend({
-  id: z.number(),
-  emailBackup: z.string().nullable(),
-  emailStripe: z.string().optional(),
-  nameFirst: z.string().nullable(),
-  nameLast: z.string().nullable(),
-  profileBlurb: z.string().nullable(),
-  profileContactEmail: z.string().nullable(),
-  profileCurrentJobCompany: z.string().nullable(),
-  profileCurrentJobTitle: z.string().nullable(),
-  profileGitHubUri: z.string().nullable(),
-  profileHighestDegree: z.string().nullable(),
-  profileHomepageUri: z.string().nullable(),
-  profileLinkedInUri: z.string().nullable(),
-  profileTopNetworkingReasons: z.array(z.string()).nullable(),
-  profileTopServices: z.array(z.string()).nullable(),
-  profileTopSkills: z.array(z.string()).nullable(),
-  profileYearsOfExperience: z.number().nullable(),
-  subscription: z.object({
-    tier: z.nativeEnum(PaymentTierEnum),
-    type: z.string(),
-  }),
-})
-
-export type UserSettings = z.infer<typeof GetUserSettingsSchema>
-export type UserWithSubscriptions = User & { subscriptions: Subscription[] }
-export type UserWithSubscriptionsOrZero =
-  | UserWithSubscriptions
-  | typeof NULL_RESULT_TRPC_INT
+import {
+  GetUserSettingsSchema,
+  UpdateUserSettingsSchema,
+  UserSettings,
+  UserWithSubscriptions,
+  UserWithSubscriptionsOrZero,
+} from '~/server/schemas'
+import { tiersOrder } from '~/server/schemas'
 
 export const userRouter = createTRPCRouter({
   getCurrentUser: publicProcedure.query(
@@ -407,8 +343,8 @@ export const userRouter = createTRPCRouter({
           profileHighestDegree: input.profileHighestDegree?.trim() ?? undefined,
           profileHomepageUri: input.profileHomepageUri?.trim() ?? null,
           profileLinkedInUri: input.profileLinkedInUri?.trim() ?? null,
-          profileTopNetworkingReasons: input.profileTopNetworkingReasons,
-          profileTopSkills: input.profileTopSkills,
+          // profileTopNetworkingReasons: input.profileTopNetworkingReasons,
+          // profileTopSkills: input.profileTopSkills,
           profileYearsOfExperience: input.profileYearsOfExperience,
           residenceCountry: input.residenceCountry?.trim() ?? '',
           residenceUSState: input.residenceUSState?.trim() ?? '',
